change controller & service returned values, exceptions now extends runtimeEx, add ErrorHandler

+++ src/main/java/ru/aston/module4/springRep/UserRepository.java
@@ -3,8 +3,8 @@ package ru.aston.module4.springRep;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import ru.aston.module4.model.User;
+import ru.aston.module4.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
